=LSTM with 1 or 2 layers overfits on fragment
Former-commit-id: a80479cf535d78fef3490a9bf96ce0432d3e8e24 [formerly 51b65146d50b9e601c4df4995894122615e9f57a]
Former-commit-id: 4fdaa6b4296957c5c270b19c338483b126245c16
Former-commit-id: ecaa34ef1fbdd596a406e1811f366b9c485405cb

+++ GNN/Models/MyRNN.py
@@ -25,7 +25,7 @@ class GRU_RNN(torch.nn.Module):
 
         # The embeddings matrix for: senses, globals, definitions, examples (the latter 2 may have gradient set to 0)
         self.X = Parameter(data.x.clone().detach(), requires_grad=True)
-        self.select_first_node = Parameter(torch.tensor([0]), requires_grad=False)
+        self.select_first_node = Parameter(torch.tensor([0]).to(dtype=torch.float32), requires_grad=False)
         self.embedding_zeros = Parameter(torch.zeros(size=(1, self.d)), requires_grad=False)
 
         # Input signals: current global’s word embedding || global’s node-state (|| sense’s node state)
